requirements.txt
fastapi
uvicorn
sqlalchemy
passlib
bcrypt
pymysql
python-multipart  # 上传文件
python-jose
aiofiles
requests
pytest
jinja2
websockets
pytesseract